refactor: 스플래시 스크린 노출 위치 변경
